pyproject.toml
[tool.poetry]
name = "test-task-pixelplex"
version = "0.1.0"
description = ""
authors = ["kirillstepankov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
aiohttp = "^3.9.5"
asyncpg = "^0.29.0"
black = "^24.4.2"
isort = "^5.13.2"
celery = "^5.4.0"
redis = "^5.0.6"
py-solc-x = "^2.0.3"
py-solc = "^3.2.0"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"

[tool.pytest.ini_options]
python_files = [
  'tests.py',
  'test_*.py',
  '*_tests.py'
]
pythonpath = '. app'
addopts = "-vv -x --lf"
testpaths = [
    "tests",
]
asyncio_mode='auto'
